input ok

+++ philo.h
@@ -5,14 +5,15 @@
 # include <pthread.h>
 # include <stdlib.h>
 # include <unistd.h>
+# include <limits.h>
 
 typedef struct s_data
 {
-	int		n_philos;
-	int		time_to_die;
-	int		time_to_sleep;
-	int		time_to_eat;
-	int		repeat;
+	long		n_philos;
+	long		time_to_die;
+	long		time_to_sleep;
+	long		time_to_eat;
+	long		repeat;
 }				t_data;
 
 #endif
